⚙️ config(server): refactor Inngest integration and add event schemas for user and space events
